updated java jdk 10 12 and drive api gui to v3

+++ src/main/java/com/company/DriveTableModel.java
@@ -41,7 +41,7 @@ public class DriveTableModel extends AbstractTableModel{
         }
     }
     // calling the insert values method
-    public void insertValues(String filename, double fileSize, String date, String action){
+    public void insertValues(String filename, long fileSize, String date, String action){
         db.insertValues(filename, fileSize, date, action);
         driveMap = db.getAllData();
     }
